requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydub==0.25.1
pytest==7.4.3
pytest-cov==4.1.0
